+++ Source/Core/Element.cpp
@@ -429,15 +429,13 @@ BoxArea Element::GetClipArea() const
 	return clip_area;
 }
 
-void Element::SetScrollableOverflowRectangle(Vector2f _scrollable_overflow_rectangle)
+void Element::SetScrollableOverflowRectangle(Vector2f _scrollable_overflow_rectangle, bool clamp_scroll_offset)
 {
 	if (scrollable_overflow_rectangle != _scrollable_overflow_rectangle)
 	{
 		scrollable_overflow_rectangle = _scrollable_overflow_rectangle;
-
-		scroll_offset.x = Math::Min(scroll_offset.x, GetScrollWidth() - GetClientWidth());
-		scroll_offset.y = Math::Min(scroll_offset.y, GetScrollHeight() - GetClientHeight());
-		DirtyAbsoluteOffset();
+		if (clamp_scroll_offset)
+			ClampScrollOffset();
 	}
 }
 
@@ -2905,6 +2903,28 @@ void Element::DirtyFontFaceRecursive()
 	const int num_children = GetNumChildren(true);
 	for (int i = 0; i < num_children; ++i)
 		GetChild(i)->DirtyFontFaceRecursive();
+}
+
+void Element::ClampScrollOffset()
+{
+	const Vector2f new_scroll_offset = {
+		Math::Min(scroll_offset.x, GetScrollWidth() - GetClientWidth()),
+		Math::Min(scroll_offset.y, GetScrollHeight() - GetClientHeight()),
+	};
+
+	if (new_scroll_offset != scroll_offset)
+	{
+		scroll_offset = new_scroll_offset;
+		DirtyAbsoluteOffset();
+	}
+}
+
+void Element::ClampScrollOffsetRecursive()
+{
+	ClampScrollOffset();
+	const int num_children = GetNumChildren(true);
+	for (int i = 0; i < num_children; ++i)
+		GetChild(i)->ClampScrollOffsetRecursive();
 }
 
 } // namespace Rml